requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
filelock==3.12.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
Pillow==9.5.0
requests==2.31.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tqdm @ git+https://github.com/tqdm/tqdm.git@5587f0d9602c0a75367ae652b2f283323f3a31ab
typing_extensions==4.6.3
urllib3==1.26.6
